Backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
sqlparse==0.5.5
tzdata==2025.3
